fix: solving linter warnings
