+++ lib/components/secondary_button.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+
+class SecondaryButton extends StatelessWidget {
+  const SecondaryButton(this.buttonText, this.onPress);
+
+  final Function onPress;
+  final String buttonText;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 40,
+      alignment: Alignment.centerLeft,
+      decoration: const BoxDecoration(
+        gradient: LinearGradient(
+          colors: <Color>[
+            Color(0xFF66BB6A),
+            Color(0xFF388E3C),
+            Color(0xFF1B5E20),
+          ],
+        ),
+        borderRadius: BorderRadius.all(
+          Radius.circular(5),
+        ),
+      ),
+      child: SizedBox.expand(
+        child: FlatButton(
+          onPressed: onPress,
+          child: Text(
+            buttonText,
+            style: const TextStyle(
+              fontSize: 20,
+              fontWeight: FontWeight.bold,
+              color: Colors.white,
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
