+++ lib/pages/Statistics/Tabs/PointsPerDayChart.dart
@@ -1,5 +1,5 @@
 import 'package:bizinuca/components/Menu.dart';
-import 'package:bizinuca/components/PrimaryButton.dart';
+import 'package:bizinuca/components/primary_button.dart';
 import 'package:bizinuca/models/PointsPerDay.dart';
 import 'package:bizinuca/pages/Statistics/CustomWidgets/ChartInLine.dart';
 import 'package:flutter/material.dart';
